feat: replace codemirror with monaco editor
