+++ packages/rath-client/package.json
@@ -11,7 +11,6 @@
   "dependencies": {
     "@antv/g6": "^4.7.10",
     "@babel/parser": "^7.19.6",
-    "@codemirror/lang-sql": "^6.3.0",
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
     "@fluentui/font-icons-mdl2": "^8.5.4",
@@ -26,7 +25,6 @@
     "airtable": "^0.11.4",
     "ali-react-table": "^2.6.1",
     "buffer": "^6.0.3",
-    "codemirror": "^6.0.1",
     "dayjs": "^1.11.6",
     "immer": "^9.0.6",
     "jsesc": "^3.0.2",
