+++ packages/rath-client/src/components/dbGraph/config.ts
@@ -1,0 +1,5 @@
+export const STROKE_RADIUS = 12;
+export const BOX_WIDTH = 120;
+export const BOX_HEIGHT = 40;
+export const GRAPH_HEIGHT = 400;
+export const BOX_PADDING = 20;
