+++ packages/rath-client/src/components/dbGraph/utils.ts
@@ -1,4 +1,5 @@
 import type { TableInfo } from "../../pages/dataSource/selection/database/api";
+import { BOX_HEIGHT, BOX_WIDTH, STROKE_RADIUS } from "./config";
 import type { IDBEdge, IDBGraph } from "./localTypes";
 
 
@@ -63,4 +64,55 @@ export const hasCircle = (edges: Readonly<IDBEdge[]>, from: number, to: number):
     };
 
     return walk(from);
+};
+
+export const encodePath = (x1: number, y1: number, x2: number, y2: number, isPreview = false): string => {
+    const main = (() => {
+        if (Math.abs(x1 - x2) < 1.2 * BOX_WIDTH) {
+            const cy = (y1 + y2) / 2;
+            const ya = cy - Math.sign(y2 - y1) * STROKE_RADIUS;
+            const xb = x1 + Math.sign(x2 - x1) * STROKE_RADIUS;
+            const xc = x2 - Math.sign(x2 - x1) * STROKE_RADIUS;
+            const yd = cy + Math.sign(y2 - y1) * STROKE_RADIUS;
+            if (Math.abs(x1 - x2) < 2 * STROKE_RADIUS) {
+                return `M${x1},${y1} V${ya} C${x1},${cy} ${x2},${cy} ${x2},${yd} V${y2}`;
+            }
+            return `M${x1},${y1} V${ya} Q${x1},${cy} ${xb},${cy} H${xc} Q${x2},${cy} ${x2},${yd} V${y2}`;
+        } else {
+            const cx = (x1 + x2) / 2;
+            const xa = cx - Math.sign(x2 - x1) * STROKE_RADIUS;
+            const yb = y1 + Math.sign(y2 - y1) * STROKE_RADIUS;
+            const yc = y2 - Math.sign(y2 - y1) * STROKE_RADIUS;
+            const xd = cx + Math.sign(x2 - x1) * STROKE_RADIUS;
+            if (Math.abs(y1 - y2) < 2 * STROKE_RADIUS) {
+                return `M${x1},${y1} H${xa} C${cx},${y1} ${cx},${y2} ${xd},${y2} H${x2}`;
+            }
+            return `M${x1},${y1} H${xa} Q${cx},${y1} ${cx},${yb} V${yc} Q${cx},${y2} ${xd},${y2} H${x2}`;
+        }
+    })();
+
+    const arrow = (() => {
+        if (isPreview) {
+            return '';
+        }
+        if (Math.abs(x1 - x2) < 1.2 * BOX_WIDTH) {
+            const x = x2;
+            if (y1 >= y2) {
+                const y = y2 + BOX_HEIGHT / 2;
+                return `M${x},${y} l-6,6 M${x},${y} l6,6`;
+            }
+            const y = y2 - BOX_HEIGHT / 2;
+            return `M${x},${y} l-6,-6 M${x},${y} l6,-6`;
+        } else {
+            const y = y2;
+            if (x1 >= x2) {
+                const x = x2 + BOX_WIDTH / 2;
+                return `M${x},${y} l6,-6 M${x},${y} l6,6`;
+            }
+            const x = x2 - BOX_WIDTH / 2;
+            return `M${x},${y} l-6,-6 M${x},${y} l-6,6`;
+        }
+    })();
+
+    return `${main}${arrow}`;
 };
