+++ packages/rath-client/src/pages/dataSource/selection/database/query-editor/interfaces.ts
@@ -1,0 +1,9 @@
+import { TableInfo } from "../api";
+
+export interface QueryEditorProps {
+    tableEnumerable: boolean;
+    tables: TableInfo[];
+    query: string;
+    setQuery: (query: string) => void;
+    preview: (query?: string) => void;
+}
