+++ packages/rath-client/src/utils/request.ts
@@ -1,5 +1,13 @@
 import { IResponse, IRow } from '../interfaces';
 
+export function errorCodeHandler (res: Response) {
+    if (res.status === 200) return;
+    if (res.status === 500) return;
+    if (res.status === 404) {
+        throw new Error('Fail to connect the server, check your network.')
+    }
+    throw new Error(res.statusText)
+}
 async function getRequest<T extends IRow = IRow, R = void>(path: string, payload?: T): Promise<R> {
     const url = new URL(path);
     if (payload) {
@@ -10,6 +18,7 @@ async function getRequest<T extends IRow = IRow, R = void>(path: string, payload
     const res = await fetch(url.toString(), {
         credentials: 'include',
     });
+    errorCodeHandler(res);
     const result = (await res.json()) as IResponse<R>;
     if (result.success) {
         return result.data;
@@ -28,6 +37,7 @@ async function postRequest<T extends IRow = IRow, R = void>(path: string, payloa
         },
         body: JSON.stringify(payload),
     });
+    errorCodeHandler(res);
     const result = (await res.json()) as IResponse<R>;
     if (result.success) {
         return result.data;
